[#113] Move Hono initialization logic out of SpringBoot Application class into hono-server module
